move to lines_4000_5000 file
