+++ src/compiler/factory/node_factory/lines_3000_4000.rs
@@ -887,42 +887,4 @@ impl<TBaseNodeFactory: 'static + BaseNodeFactory> NodeFactory<TBaseNodeFactory> 
         );
         node
     }
-
-    pub fn create_assert_clause(
-        &self,
-        base_factory: &TBaseNodeFactory,
-        elements: NodeArray, /*<AssertEntry>*/
-        multi_line: Option<bool>,
-    ) -> AssertClause {
-        let node = self.create_base_node(base_factory, SyntaxKind::AssertClause);
-        let mut node = AssertClause::new(node, elements, multi_line);
-        node.add_transform_flags(TransformFlags::ContainsESNext);
-        node
-    }
-
-    pub fn create_assert_entry(
-        &self,
-        base_factory: &TBaseNodeFactory,
-        name: Rc<Node /*AssertionKey*/>,
-        value: Rc<Node /*StringLiteral*/>,
-    ) -> AssertEntry {
-        let node = self.create_base_node(base_factory, SyntaxKind::AssertEntry);
-        let mut node = AssertEntry::new(node, name, value);
-        node.add_transform_flags(TransformFlags::ContainsESNext);
-        node
-    }
-
-    pub fn create_namespace_import(
-        &self,
-        base_factory: &TBaseNodeFactory,
-        name: Rc<Node /*Identifier*/>,
-    ) -> NamespaceImport {
-        let node = self.create_base_node(base_factory, SyntaxKind::NamespaceImport);
-        let mut node = NamespaceImport::new(node, name);
-        node.add_transform_flags(propagate_child_flags(Some(&*node.name)));
-        node.set_transform_flags(
-            node.transform_flags() & !TransformFlags::ContainsPossibleTopLevelAwait,
-        );
-        node
-    }
 }
