+++ src/compiler/factory/node_factory/lines_4000_5000.rs
@@ -1,12 +1,51 @@
 use std::rc::Rc;
 
-use super::get_default_tag_name_for_kind;
+use super::{get_default_tag_name_for_kind, propagate_child_flags};
 use crate::{
-    BaseJSDocTag, BaseJSDocTypeLikeTag, BaseJSDocUnaryType, BaseNode, BaseNodeFactory, JsxText,
-    Node, NodeArray, NodeFactory, NodeInterface, StringOrNodeArray, SyntaxKind, TransformFlags,
+    AssertClause, AssertEntry, BaseJSDocTag, BaseJSDocTypeLikeTag, BaseJSDocUnaryType, BaseNode,
+    BaseNodeFactory, JsxText, NamespaceImport, Node, NodeArray, NodeFactory, NodeInterface,
+    StringOrNodeArray, SyntaxKind, TransformFlags,
 };
 
 impl<TBaseNodeFactory: 'static + BaseNodeFactory> NodeFactory<TBaseNodeFactory> {
+    pub fn create_assert_clause(
+        &self,
+        base_factory: &TBaseNodeFactory,
+        elements: NodeArray, /*<AssertEntry>*/
+        multi_line: Option<bool>,
+    ) -> AssertClause {
+        let node = self.create_base_node(base_factory, SyntaxKind::AssertClause);
+        let mut node = AssertClause::new(node, elements, multi_line);
+        node.add_transform_flags(TransformFlags::ContainsESNext);
+        node
+    }
+
+    pub fn create_assert_entry(
+        &self,
+        base_factory: &TBaseNodeFactory,
+        name: Rc<Node /*AssertionKey*/>,
+        value: Rc<Node /*StringLiteral*/>,
+    ) -> AssertEntry {
+        let node = self.create_base_node(base_factory, SyntaxKind::AssertEntry);
+        let mut node = AssertEntry::new(node, name, value);
+        node.add_transform_flags(TransformFlags::ContainsESNext);
+        node
+    }
+
+    pub fn create_namespace_import(
+        &self,
+        base_factory: &TBaseNodeFactory,
+        name: Rc<Node /*Identifier*/>,
+    ) -> NamespaceImport {
+        let node = self.create_base_node(base_factory, SyntaxKind::NamespaceImport);
+        let mut node = NamespaceImport::new(node, name);
+        node.add_transform_flags(propagate_child_flags(Some(&*node.name)));
+        node.set_transform_flags(
+            node.transform_flags() & !TransformFlags::ContainsPossibleTopLevelAwait,
+        );
+        node
+    }
+
     pub(crate) fn create_jsdoc_primary_type_worker(
         &self,
         base_factory: &TBaseNodeFactory,
